+++ tests/test_gematria.py
@@ -52,3 +52,17 @@ def test_mispar_hechrachi():
         ).gematria(GematriaTypes.MISPAR_HECHRACHI)
         == 2701
     )
+
+
+def test_bad_input():
+    for t in GematriaTypes:
+        assert (
+            Hebrew("Testing a string that contains no hebrew").gematria(t) == 0
+        ), "gematria should return 0 when the input text is not hebrew"
+
+
+def test_mixed_input():
+    for t in GematriaTypes:
+        assert (
+            Hebrew("Both hebrew: 'שָׁלוֹם' and English: 'Hello dev!'").gematria(t) > 0
+        ), "gematria should always have a value > 0 when there is any hebrew text in a string"