+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.4] - 2021-11-21
+
+### Fixed
+
+- Fixed an issue where `Hebrew.gematria` would through an error if the input string had no hebrew characters.
+  In this case, we now return a value of 0.
+
 ## [0.5.3] - 2021-11-15
 
 ### Changed